Automatically copy the MSW service worker after a dependency update
This is required as part of the MSW installation, so this will avoid having to manually do it when Renovate opens a PR updating this package. I have not found a way to run this only if the `msw` package is updated, but this should be a no-op in the other cases.