[Vite] Remove overridden build.target in favor of legacy plugin defaults
With the target option specified, a warning that such is being overridden is raised by the legacy plugin:

```
plugin-legacy overrode 'build.target'. You should pass 'modernTargets' as an option to this plugin with the list of browsers to support instead.
```

With it removed, essentially allowing the legacy plugin to determine the targets via its defaults (we can specify if we want), the warning no longer appears.