Change display of blocked and muted quoted users
Instead of hiding the content with “Post unavailable”, this shows whether you
have muted or blocked the account, or the account's domain, and allows you to
show the post anyway.

This is a significant change to the API, returning the status in those cases
whereas it was previously omitted. It also adds `blocked_domain`, `blocked_account`
and `muted_account` as `state` values.