Fix creation of duplicate conversations
And make sure the conversation that is created is wired up
correctly.

Before this change, the cyclical relationship between status and
conversation would lead to two records being created, one with
the correct `parent_status_id` and one without. For unknown
reasons, the one without ended up being the status' conversation
in production environments (but never in tests).