Refactor relation preloading
Preloading relations is a massive win in terms of preventing
superfluous db queries. But the complex logic needed to use this
was copied all over the place.

This change attempts to encapsulate the logic in the account
itself, thus simplifying some areas that used it. It should also
prevent future copy&pasting of this logic.